🔵 Move loadFixture to TestCase
