+++ src/integrationTest/kotlin/com/capitole/capitoleproductcatalogapi/infrastructure/testcases/TestCase.kt
@@ -2,6 +2,8 @@ package com.capitole.capitoleproductcatalogapi.infrastructure.testcases
 
 
 import com.capitole.capitoleproductcatalogapi.Application
+import com.fasterxml.jackson.databind.JsonNode
+import com.fasterxml.jackson.databind.ObjectMapper
 import io.restassured.module.mockmvc.RestAssuredMockMvc.mockMvc
 import org.junit.jupiter.api.BeforeEach
 import org.springframework.beans.factory.annotation.Autowired
@@ -9,6 +11,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.web.servlet.MockMvc
+import java.nio.charset.StandardCharsets
 
 @SpringBootTest(classes = [Application::class])
 @ActiveProfiles("integration-test")
@@ -17,8 +20,18 @@ abstract class TestCase {
 
   @Autowired private lateinit var mvc: MockMvc
 
+  protected lateinit var mapper: ObjectMapper
+
   @BeforeEach
   fun setUp() {
+    mapper = ObjectMapper()
     mockMvc(mvc)
   }
+
+  protected fun loadFixture(name: String): JsonNode =
+      javaClass.classLoader
+          .getResourceAsStream("responses/$name")
+          ?.bufferedReader(StandardCharsets.UTF_8)
+          ?.use { mapper.readTree(it) }
+        ?: error("Fixture not found: responses/$name")
 }
